#0: refactor ttnn and tracy module   : update fmt for ttnn/unary.py and use __all__ to be list of symbol strings   : reorg / trim imports; import times remain same as with sys.module / attr hack
