+++ ttnn/decorators.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from functools import wraps
+
 from loguru import logger
